requirements.txt
# Flask Framework
Flask==3.0.0
Flask-Cors==4.0.0

# Database Connectors
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Security & Authentication
bcrypt==4.1.1
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time handling
python-dateutil==2.8.2

# Development Tools
Werkzeug==3.0.1
